+++ baseapp/baseapp_test.go
@@ -912,194 +912,6 @@ func TestConcurrentCheckDeliver(t *testing.T) {
 	// TODO
 }
 
-// Test that transactions exceeding gas limits fail
-func TestTxGasLimitsOld(t *testing.T) {
-	gasGranted := uint64(10)
-	anteOpt := func(bapp *BaseApp) {
-		bapp.SetAnteHandler(func(ctx sdk.Context, tx sdk.Tx, simulate bool) (newCtx sdk.Context, err error) {
-			newCtx = ctx.WithGasMeter(sdk.NewGasMeter(gasGranted))
-
-			// AnteHandlers must have their own defer/recover in order for the BaseApp
-			// to know how much gas was used! This is because the GasMeter is created in
-			// the AnteHandler, but if it panics the context won't be set properly in
-			// runTx's recover call.
-			defer func() {
-				if r := recover(); r != nil {
-					switch rType := r.(type) {
-					case sdk.ErrorOutOfGas:
-						err = sdkerrors.Wrapf(sdkerrors.ErrOutOfGas, "out of gas in location: %v", rType.Descriptor)
-					default:
-						panic(r)
-					}
-				}
-			}()
-
-			count := tx.(txTest).Counter
-			newCtx.GasMeter().ConsumeGas(uint64(count), "counter-ante")
-
-			return newCtx, nil
-		})
-	}
-
-	routerOpt := func(bapp *BaseApp) {
-		// r := sdk.NewRoute(routeMsgCounter, func(ctx sdk.Context, msg sdk.Msg) (*sdk.Result, error) {
-		// 	count := msg.(*msgCounter).Counter
-		// 	ctx.GasMeter().ConsumeGas(uint64(count), "counter-handler")
-		// 	return &sdk.Result{}, nil
-		// })
-		// bapp.Router().AddRoute(r)
-	}
-
-	app := setupBaseApp(t, anteOpt, routerOpt)
-
-	header := tmproto.Header{Height: 1}
-	app.BeginBlock(abci.RequestBeginBlock{Header: header})
-
-	testCases := []struct {
-		tx      *txTest
-		gasUsed uint64
-		fail    bool
-	}{
-		{newTxCounter(0, 0), 0, false},
-		{newTxCounter(1, 1), 2, false},
-		{newTxCounter(9, 1), 10, false},
-		{newTxCounter(1, 9), 10, false},
-		{newTxCounter(10, 0), 10, false},
-		{newTxCounter(0, 10), 10, false},
-		{newTxCounter(0, 8, 2), 10, false},
-		{newTxCounter(0, 5, 1, 1, 1, 1, 1), 10, false},
-		{newTxCounter(0, 5, 1, 1, 1, 1), 9, false},
-
-		{newTxCounter(9, 2), 11, true},
-		{newTxCounter(2, 9), 11, true},
-		{newTxCounter(9, 1, 1), 11, true},
-		{newTxCounter(1, 8, 1, 1), 11, true},
-		{newTxCounter(11, 0), 11, true},
-		{newTxCounter(0, 11), 11, true},
-		{newTxCounter(0, 5, 11), 16, true},
-	}
-
-	for i, tc := range testCases {
-		tx := tc.tx
-		gInfo, result, err := app.SimDeliver(aminoTxEncoder(), tx)
-
-		// check gas used and wanted
-		require.Equal(t, tc.gasUsed, gInfo.GasUsed, fmt.Sprintf("tc #%d; gas: %v, result: %v, err: %s", i, gInfo, result, err))
-
-		// check for out of gas
-		if !tc.fail {
-			require.NotNil(t, result, fmt.Sprintf("%d: %v, %v", i, tc, err))
-		} else {
-			require.Error(t, err)
-			require.Nil(t, result)
-
-			space, code, _ := sdkerrors.ABCIInfo(err, false)
-			require.EqualValues(t, sdkerrors.ErrOutOfGas.Codespace(), space, err)
-			require.EqualValues(t, sdkerrors.ErrOutOfGas.ABCICode(), code, err)
-		}
-	}
-}
-
-// Test that transactions exceeding gas limits fail
-func TestMaxBlockGasLimits(t *testing.T) {
-	gasGranted := uint64(10)
-	anteOpt := func(bapp *BaseApp) {
-		bapp.SetAnteHandler(func(ctx sdk.Context, tx sdk.Tx, simulate bool) (newCtx sdk.Context, err error) {
-			newCtx = ctx.WithGasMeter(sdk.NewGasMeter(gasGranted))
-
-			defer func() {
-				if r := recover(); r != nil {
-					switch rType := r.(type) {
-					case sdk.ErrorOutOfGas:
-						err = sdkerrors.Wrapf(sdkerrors.ErrOutOfGas, "out of gas in location: %v", rType.Descriptor)
-					default:
-						panic(r)
-					}
-				}
-			}()
-
-			count := tx.(txTest).Counter
-			newCtx.GasMeter().ConsumeGas(uint64(count), "counter-ante")
-
-			return
-		})
-	}
-
-	routerOpt := func(bapp *BaseApp) {
-		// r := sdk.NewRoute(routeMsgCounter, func(ctx sdk.Context, msg sdk.Msg) (*sdk.Result, error) {
-		// 	count := msg.(*msgCounter).Counter
-		// 	ctx.GasMeter().ConsumeGas(uint64(count), "counter-handler")
-		// 	return &sdk.Result{}, nil
-		// })
-		// bapp.Router().AddRoute(r)
-	}
-
-	app := setupBaseApp(t, anteOpt, routerOpt)
-	app.InitChain(abci.RequestInitChain{
-		ConsensusParams: &abci.ConsensusParams{
-			Block: &abci.BlockParams{
-				MaxGas: 100,
-			},
-		},
-	})
-
-	testCases := []struct {
-		tx                *txTest
-		numDelivers       int
-		gasUsedPerDeliver uint64
-		fail              bool
-		failAfterDeliver  int
-	}{
-		{newTxCounter(0, 0), 0, 0, false, 0},
-		{newTxCounter(9, 1), 2, 10, false, 0},
-		{newTxCounter(10, 0), 3, 10, false, 0},
-		{newTxCounter(10, 0), 10, 10, false, 0},
-		{newTxCounter(2, 7), 11, 9, false, 0},
-		{newTxCounter(10, 0), 10, 10, false, 0}, // hit the limit but pass
-
-		{newTxCounter(10, 0), 11, 10, true, 10},
-		{newTxCounter(10, 0), 15, 10, true, 10},
-		{newTxCounter(9, 0), 12, 9, true, 11}, // fly past the limit
-	}
-
-	for i, tc := range testCases {
-		tx := tc.tx
-
-		// reset the block gas
-		header := tmproto.Header{Height: app.LastBlockHeight() + 1}
-		app.BeginBlock(abci.RequestBeginBlock{Header: header})
-
-		// execute the transaction multiple times
-		for j := 0; j < tc.numDelivers; j++ {
-			_, result, err := app.SimDeliver(aminoTxEncoder(), tx)
-
-			ctx := app.getState(runTxModeDeliver).ctx
-
-			// check for failed transactions
-			if tc.fail && (j+1) > tc.failAfterDeliver {
-				require.Error(t, err, fmt.Sprintf("tc #%d; result: %v, err: %s", i, result, err))
-				require.Nil(t, result, fmt.Sprintf("tc #%d; result: %v, err: %s", i, result, err))
-
-				space, code, _ := sdkerrors.ABCIInfo(err, false)
-				require.EqualValues(t, sdkerrors.ErrOutOfGas.Codespace(), space, err)
-				require.EqualValues(t, sdkerrors.ErrOutOfGas.ABCICode(), code, err)
-				require.True(t, ctx.BlockGasMeter().IsOutOfGas())
-			} else {
-				// check gas used and wanted
-				blockGasUsed := ctx.BlockGasMeter().GasConsumed()
-				expBlockGasUsed := tc.gasUsedPerDeliver * uint64(j+1)
-				require.Equal(
-					t, expBlockGasUsed, blockGasUsed,
-					fmt.Sprintf("%d,%d: %v, %v, %v, %v", i, j, tc, expBlockGasUsed, blockGasUsed, result),
-				)
-
-				require.NotNil(t, result, fmt.Sprintf("tc #%d; currDeliver: %d, result: %v, err: %s", i, j, result, err))
-				require.False(t, ctx.BlockGasMeter().IsPastLimit())
-			}
-		}
-	}
-}
-
 // Test custom panic handling within app.DeliverTx method
 func TestCustomRunTxPanicHandler(t *testing.T) {
 	const customPanicMsg = "test panic"
